Inventory fix and output upgrade
Added functionality so that events which do not pair with a schedule will appear on the front page indicating a new schedule needs to be made for that day. Colored these cells green.

Fixed a bug where days would not appear becuase the date capturing code in parse_siemens_schedule would catch too many items.