Fix many to 1 schedules
Event facilities that require multiple schedules will not be handled properly. The solution was to split the "Siemens Schedule" values by ',' and duplicate all of the other row information so that the matching apparatus still works properly. This just creates a redundancy under the "Who booked this room" section, but that is not a true issue.

Updated Everson Class Schedule to be "Can not schedule" Updated MU S8 to be inside of the tower group.